Back to commit #31a0e8f
Tried several different ways to add reposts and duplicate control in a way that keeps the efficiency of the system the same, but failed. A new strategy will be developed for this. Model cloning is planned for each version of the notes.

Let's say the main/original note A is added to the feed.
A repost of this note, B, is not currently showing in the feed.
But unnecessarily reprocessing these kind 6 events is not a good choice for resource consumption.

So B will be added to the feed only by cloning A (plus repost data). This way there will be no separate data fetching and processing for B.